[♻️refactor]: ContextMenu의 interface들 분리
